Backend/requirements.txt
Flask
flask-cors
python-dotenv
groq
pinecone>=3.0.0
requests
